Patch meta fixes
Fixed a bug with patch meta keys (_pm, _ps) not working on func_group.

Patch subdivisions now can be a float value (0.5, 0.25 and so on).

+++ src/mesh.c
@@ -520,7 +520,7 @@ given a curve of a certain length, return the number of subdivision iterations
 note: this is affected by subdivision amount
 */
 
-int IterationsForCurve( float len, int subdivisions )
+int IterationsForCurve( float len, float subdivisions )
 {
 	int		iterations, facets;
